blogging_api/requirements.txt
Django~=4.2.11

djangorestframework

djangorestframework-simplejwt-5.3.1 pyjwt-2.9.0

django-filter